+++ Assets/FluentMotion/ReactiveBehaviour.cs
@@ -1,0 +1,12 @@
+using System;
+using FluentMotion.hand;
+using Leap;
+using UnityEngine;
+
+namespace FluentMotion
+{
+    public class ReactiveHandBehaviour : MonoBehaviour
+    {
+        protected IObservable<Hand> Hand => gameObject.GetComponent<ReactiveHand>().AsObservable();
+    }
+}