+++ Assets/FluentMotion/ReactiveHandsBehaviour.cs
@@ -1,0 +1,13 @@
+using System;
+using FluentMotion.hand;
+using FluentMotion.helpers;
+using Leap;
+using UnityEngine;
+
+namespace FluentMotion
+{
+    public class ReactiveHandsBehaviour : MonoBehaviour
+    {
+        protected IObservable<ReactiveHandsHelper<Hand>> Hands => gameObject.GetComponent<ReactiveHands>().AsObservable();
+    }
+}