+++ Assets/FluentMotion/finger/impl/Pinky.cs
@@ -1,0 +1,10 @@
+using Leap;
+using Leap.Unity;
+
+namespace FluentMotion.finger.impl
+{
+    public class Pinky
+    {
+        public static bool IsExtended(Hand hand) => hand.GetPinky().IsExtended;
+    }
+}