+++ Assets/FluentMotion/finger/impl/Ring.cs
@@ -1,0 +1,10 @@
+using Leap;
+using Leap.Unity;
+
+namespace FluentMotion.finger.impl
+{
+    public class Ring
+    {
+        public static bool IsExtended(Hand hand) => hand.GetRing().IsExtended;
+    }
+}