+++ Assets/FluentMotion/finger/impl/Thumb.cs
@@ -1,4 +1,6 @@
-﻿using Leap;
+﻿using System;
+using FluentMotion.helpers.hand;
+using Leap;
 using Leap.Unity;
 
 namespace FluentMotion.finger.impl
@@ -6,6 +8,5 @@ namespace FluentMotion.finger.impl
     public static class Thumb
     {
         public static bool IsExtended(Hand hand) => hand.GetThumb().IsExtended;
-        public static bool IsExtended(Finger finger) => finger.IsExtended;
     }
 }